+++ Inc/uart_ll.h
@@ -6,7 +6,7 @@
 
 #include "ring_buffer.h"
 
-void uart_ll_init(size_t baud, ring_buffer *rx_buf, ring_buffer *tx_buf);
+void uart_ll_init(uint32_t baud, ring_buffer *rx_buf, ring_buffer *tx_buf);
 void uart_ll_start();
 void uart_ll_transmit(uint8_t data);
